BldSql is created

+++ region/dspDlt.php
@@ -5,8 +5,11 @@
  * Date: 19/6/2015
  * Time: 0:16
  */
+$regCd = @$HTTP_RAW_POST_DATA;
+if (is_null($regCd)) exit();
+echo "regCd=" . $regCd;
 include_once '/../phpFn/db.php';
 $con = db_con();
-$regCd = $_GET['regCd'];
 runsql_exec($con, "delete from region where regCd= '$regCd'; ");
 $con->close();
+
